Updated: code to add animation creation logic to root level and removed it from component level(#23zj3yc)
