+++ src/views/Product.vue
@@ -8,7 +8,7 @@
     </ion-header>
 
     <ion-content>
-      <main ref="main">
+      <main>
         <section>
           <ion-card>
             <ion-card-header>
@@ -28,7 +28,7 @@
           </ion-card>
         </section>
 
-        <aside class="desktop-only" v-show="currentJob" ref="aside">
+        <aside class="desktop-only" v-show="currentJob">
           <JobDetail :title="title" :job="currentJob" :status="currentJobStatus" :type="freqType" :key="currentJob"/>
         </aside>
       </main>
@@ -38,7 +38,6 @@
 
 <script lang="ts">
 import {
-  createAnimation,
   IonCard,
   IonCardHeader,
   IonCardTitle,
@@ -51,9 +50,10 @@ import {
   IonTitle,
   IonToolbar
 } from '@ionic/vue';
-import { defineComponent, ref } from 'vue';
+import { defineComponent } from 'vue';
 import { mapGetters, useStore } from 'vuex';
 import JobDetail from '@/components/JobDetail.vue'
+import emitter from '@/event-bus';
 
 export default defineComponent({
   name: 'Product',
@@ -105,25 +105,7 @@ export default defineComponent({
       this.freqType = this.jobFrequencyType[id]
 
       if (this.currentJob && !this.isJobDetailAnimationCompleted) {
-        const revealAnimation = createAnimation()
-        .addElement(this.aside)
-        .duration(1500)
-        .easing('ease')
-         .keyframes([
-          { offset: 0, flex: '0', opacity: '0' },
-          { offset: 0.5, flex: '1', opacity: '0' },
-          { offset: 1, flex: '1', opacity: '1' }
-        ])
-
-        const gapAnimation = createAnimation()
-          .addElement(this.main)
-          .duration(500)
-          .fromTo('gap', '0', 'var(--spacer-2xl)');
-
-        createAnimation()
-          .addAnimation([gapAnimation, revealAnimation])
-          .play();
-
+        emitter.emit('playAnimation');
         this.isJobDetailAnimationCompleted = true;
       }
     },
@@ -139,12 +121,8 @@ export default defineComponent({
       showBackdrop: false
     }
     const store = useStore();
-    const main = ref({} as Element)
-    const aside = ref({} as Element)
     return {
-      aside,
       customPopoverOptions,
-      main,
       store
     }
   }
